requirements.txt
google-search-results~=2.4.2
django~=4.2.10
pyfunctional~=1.4.3
python-dotenv~=1.0.1
brotli~=1.1.0
pandas~=2.2.0
openai~=1.12.0
Pillow~=10.2.0
django-cors-headers~=4.3.1
PyPDF2~=3.0.1
pdfplumber~=0.11.0
dill~=0.3.8
python-dateutil~=2.8.2
gunicorn~=22.0.0
dj-config-url~=0.1.1
psycopg2-binary~=2.9.9
